Updated Sales Return Form

+++ SamyuktaManabSoftwares/GrocerySupplyManagementApp/ViewModels/SalesReturnTransactionView.cs
@@ -7,7 +7,7 @@
         public string Description { get; set; }
         public string ItemCode { get; set; }
         public string ItemName { get; set; }
-        public int ItemQuantity { get; set; }
+        public decimal ItemQuantity { get; set; }
         public decimal ItemPrice { get; set; }
         public decimal SalesProfit { get; set; }
         public decimal Amount { get; set; }
